+++ utils/dataTestIds.ts
@@ -1,0 +1,7 @@
+export const DATA_TEST_IDS = {
+  hero: {
+    linkLinkedIn: 'hero-link-linkedin',
+    linkResume: 'hero-link-resume',
+    linkGitHub: 'hero-link-github',
+  },
+}
