--- utils/dataTestid.ts
@@ -1,6 +1,0 @@
-// TODO: fix file name during refactor
-export const DATA_TEST_IDS = {
-  heroLinkLinkedIn: 'hero-link-linkedin',
-  heroLinkResume: 'hero-link-resume',
-  heroLinkGitHub: 'hero-link-github',
-}
